Modularize imgui - split windows and ready to add into any app
