+++ examples/common/imgui_ui/BUILD.gn
@@ -1,0 +1,41 @@
+# Copyright (c) 2023 Project CHIP Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import("//build_overrides/chip.gni")
+
+import("${chip_root}/build/chip/tools.gni")
+import("${chip_root}/third_party/imgui/imgui.gni")
+
+config("imgui_ui_config") {
+  # allow including via 'imgui_ui/....'
+  include_dirs = [ "../" ]
+}
+
+static_library("imgui_ui") {
+  sources = [
+    "ui.cpp",
+    "ui.h",
+  ]
+
+  deps = [
+    "${chip_root}/examples/common/imgui_ui/windows",
+    "${chip_root}/examples/platform/linux:app-main",
+    "${chip_root}/src/lib/support",
+    "${chip_root}/third_party/imgui",
+  ]
+
+  public_configs = [
+    ":imgui_ui_config",
+    "${chip_root}/third_party/imgui:imgui_config",  # propagate enabling flag
+  ]
+}
