+++ examples/common/imgui_ui/windows/window.h
@@ -1,6 +1,6 @@
 /*
+ *
  *    Copyright (c) 2023 Project CHIP Authors
- *    All rights reserved.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -14,22 +14,34 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-
 #pragma once
 
 namespace example {
 namespace Ui {
 
-// MUST be called from within the main thread, to access CHIP global
-// data
-void Init();
+/**
+ * Represents a generic UI window for a ImGUI application.
+ *
+ * Provides callbacks of loading state, updating state and rendering the
+ * actual UI.
+ *
+ * UI rendering is expected to be done using imgui.
+ */
+class Window
+{
+public:
+    virtual ~Window() = default;
 
-// MAC requires this to be run from the main event loop. Linux supports
-// running this from a thread
-void EventLoop();
+    // State updates will run in the chip main loop
 
-// Flags the UI Event loop to stop
-void StopEventLoop();
+    virtual void LoadInitialState() {}
+    virtual void UpdateState() {}
+
+    // Render the UI
+    // MUST use Imgui rendering, generally within a Begin/End block to
+    // create a window.
+    virtual void Render() = 0;
+};
 
 } // namespace Ui
 } // namespace example
